[OT-347] [REFACTOR]: home api 리팩토링
