+++ src/entities/home/apis/index.ts
@@ -1,5 +1,5 @@
-export * from "./getTrendingList";
-export * from "./getTagsTop";
-export * from "./getHistoryList";
-export * from "./getMoodCard";
-export * from "./patchMoodHide";
+export * from "./getTrendingListApi";
+export * from "./getTagsTopApi";
+export * from "./getHistoryListApi";
+export * from "./getMoodCardApi";
+export * from "./patchMoodHideApi";
