--- src/entities/home/apis/patchMoodHide.ts
@@ -1,5 +1,0 @@
-import { api } from "@shared/api";
-
-export const patchMoodHide = async (refreshId: number) => { 
-    await api.patch(`/mood-refresh/${refreshId}/hide`, { refreshId });
-}
