+++ src/entities/home/apis/patchMoodHideApi.ts
@@ -1,0 +1,5 @@
+import { api } from "@shared/api";
+
+export const patchMoodHideApi = async (refreshId: number) => {
+  await api.patch(`/mood-refresh/${refreshId}/hide`, { refreshId });
+};
