+++ src/entities/home/hooks/useHideMood.ts
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { patchMoodHide } from "@entities/home/apis";
+import { patchMoodHideApi } from "@entities/home/apis";
 
 export const useHideMood = () => {
   const [isLoading, setIsLoading] = useState(false);
@@ -7,7 +7,7 @@ export const useHideMood = () => {
   const hideMood = async (refreshId: number) => {
     setIsLoading(true);
     try {
-      await patchMoodHide(refreshId);
+      await patchMoodHideApi(refreshId);
     } finally {
       setIsLoading(false);
     }
