+++ src/entities/home/hooks/useHistoryList.ts
@@ -1,9 +1,9 @@
 import { useQuery } from "@tanstack/react-query";
-import { GetHistoryListParams, historyListApi } from "@entities/home/apis";
+import { GetHistoryListParams, getHistoryListApi } from "@entities/home/apis";
 
 export const useHistoryList = (params: GetHistoryListParams) => {
   return useQuery({
     queryKey: ["HistoryList", params],
-    queryFn: () => historyListApi(params),
+    queryFn: () => getHistoryListApi(params),
   });
 };
